# swagger.yaml
openapi: 3.0.0
info:
  title: 'API Livros'
  description: 'Documentação da API Livros'
  version: 1.0.0
servers:
  -
    url: 'http://localhost/Api-Livros'
    description: 'Servidor local'
paths:
...
  /recuperar-senha:
    post:
      tags:
        - Senha
      summary: 'Solicita recuperação de senha'
      operationId: 6c4bbe3c4383d35ae0ddc5a28f6bc122
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  example: teste@email.com
              type: object
      responses:
        '200':
          description: 'Solicitação processada com sucesso'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: true }
                  mensagem: { type: string, example: 'Se o e-mail estiver cadastrado, você receberá instruções para redefinir sua senha.' }
                type: object
        '400':
          description: 'E-mail não informado'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: false }
                  mensagem: { type: string, example: 'E-mail é obrigatório' }
                type: object
        '500':
          description: 'Erro ao gerar token ou enviar e-mail'
          content:
            application/json:
              schema:
                properties:
                  mensagem: { type: string, example: 'Erro ao enviar e-mail' }
                type: object
  /redefinir-senha:
    post:
      tags:
        - Senha
      summary: 'Redefine a senha do usuário'
      operationId: f1050548a77cb0a6d6db0f7e4042d2cb
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - token
                - senha
              properties:
                token:
                  type: string
                  example: a1b2c3d4e5f6g7h8
                senha:
                  type: string
                  format: password
                  example: NovaSenha@123
              type: object
      responses:
        '200':
          description: 'Senha redefinida com sucesso'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: true }
                  mensagem: { type: string, example: 'Senha redefinida com sucesso' }
                type: object
        '401':
          description: 'Não autenticado'
        '500':
          description: 'Não foi possível atualizar a senha'
          content:
            application/json:
              schema:
                properties:
                  mensagem: { type: string, example: 'Nao foi possível atualizar a senha' }
                type: object
      security:
        -
          bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      bearerFormat: JWT
      scheme: bearer
...